client/pyproject.toml
[tool.poetry]
name = "rcpserv_client"
version = "1.0.0"
description = ""
authors = ["wxllow <willow@wxllow.dev>", "BennoCrafter"]
license = "MIT"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.9"
pypresence = "^4.2.1"
python-socketio = "^5.8.0"
requests = "^2.31.0"
aiohttp = "^3.8.4"
pystray = "^0.19.4"
pillow = "9.5.0"
pygobject = {version = "^3.44.1", platform = "Linux"}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
